+++ template/project/webgame/project/src/net/v2/network.js
@@ -9,6 +9,7 @@
 ************************************************************************ */
 import app from 'entity/app';
 
+import * as event from 'net/event';
 import fmtGame from 'schema/game';
 import * as vendor from 'src/vendor';