+++ template/project/webgame/project/src/scenes/mainSet.js
@@ -12,17 +12,72 @@ import app from 'entity/app';
 
 
 /**
- * 物件初始化
- * @param {Object} that
+ * 更新資源
  */
-export function normal (that) {
+export async function reload (scene) {
+  console.log('!!!!!!!! 開始處理 reload !!!!!!!!!');
 
-  let ui = app.nuts.ui;
+  let main = app.scenes.main;
+  let game = main.game;
+  let ent = main.entity;
+
+  // 指定要更新的資源
+  let newTextures = scene.textures.ui;
+  let textures    = main.textures;
+
+  textures.demo.ring = scene.textures.demo.ring;
+
+  let dest = textures.demo.button;
+  let src  = scene.textures.demo.button;
+
+  dest.auto      = src.auto;
+  dest.autoLight = src.autoLight;
+  dest.bet       = src.bet;
+  dest.leave     = src.leave;
+
+  let obj = null;
+  obj = ent.gui;
+  if (obj) {
+    obj.reload();
+  }
+
+  obj = ent.ring;
+  if (obj) {
+    obj.reload();
+  }
+
+  obj = ent.cash;
+  if (obj) {
+    await game.idle(1.0);
+    obj.x = 30;
+    obj.y = 50;
+    obj.custom.rect.width = 360;
+    obj.custom.rect.height = 100;
+    obj.custom.offset = 72;
+    obj.setTextures('numBW');
+  }
+
+  obj = ent.coin;
+  if (obj) {
+    await game.idle(1.0);
+    obj.texture = newTextures.pageInfoText.a;
+  }
+  console.log('!!!!!!!! 完成處理 reload !!!!!!!!!');
+}
+
+
+/**
+ * 物件初始化
+ * @param {Object} options
+ */
+export function normal (/* options */) {
+
+  let nuts = app.nuts;
+  let ui = nuts.ui;
   const NUM = ui.Number.NUM;
 
-  let center = that.getCenter();
+  let main = app.scenes.main;
 
-  // let textures = center.textures.demo;
   let ruleVisible = false;
   class Rule {
     constructor (obj, layer) {
@@ -39,7 +94,7 @@ export function normal (that) {
       let self = this;
       console.log(scene);
 
-      center.game.textures.ui = scene.textures.ui;
+      main.game.textures.ui = scene.textures.ui;
       self.group.reload();
     }
   }
@@ -59,7 +114,9 @@ export function normal (that) {
         if (ruleVisible) {
           create();
         } else {
-          center.rule.hide();
+          if (main.rule) {
+            main.rule.hide();
+          }
         }
       });
     },
@@ -79,18 +136,19 @@ export function normal (that) {
 
     async setLeave (obj) {
       async function leave () {
-        let sound = center.sounds.demo;
+        let sound = app.sounds.demo;
         if (sound && sound.music && sound.music.play) {
           sound.music.stop();
         }
         let scene = await import('scene/sub');
         scene.reset();
+        main.hide();
 
         await app.game.idle(0.01);
         let config = {
-          game: '{%= name %}',
-          group: '{%= group %}',
-          id: '{%= name %}',
+          game: 'z01',
+          group: 'demo',
+          id: 'z01',
           tablekey: 'abcd1234'
         };
         app.game.scene.reload(config);
@@ -101,14 +159,23 @@ export function normal (that) {
       });
     },
 
-    // 設定下注
-    setBet (/* obj */) {
+    // 投注額
+    setBet (obj) {
+      obj.setClick((/*o*/) => {
+        console.log('[投注額]');
+      });
     },
-    
+
+    // 自動
+    setAuto (obj) {
+      obj.setClick((/*o*/) => {
+        console.log('[自動]');
+      });
+    },
+
     // 設定動畫
     setAnim (obj) {
-      console.log('!!!!!!!!! setAnim !!!!!!!!!!');
-      console.log(obj);
+      console.log('設定動畫');
       obj.play();
     },
 
@@ -123,14 +190,16 @@ export function normal (that) {
     setSpine (obj) {
       obj.play();
     },
+
     createRule (obj) {
-      console.log('[create rule]');
+      console.log('[建立說明頁面]');
       console.log(obj);
 
-      center.rule = new Rule(obj, center.game.layer.foreground);
+      main.rule = new Rule(obj, main.game.layer.foreground);
     },
+
     autoonPageNumber (obj) {
-      obj.fixVal = center.decimal;
+      obj.fixVal = app.decimal;
       obj.setAnchor({
         x: 0.0,
         y: 0.0
@@ -141,7 +210,10 @@ export function normal (that) {
   };
 
   let config = {
-    set
+    set,
+    scene: main,
+    game: app.game
   };
-  that.init(config);
+
+  nuts.scene.set(config);
 }